Clean up how queries are exposed externally
Previously, the getters returned from flow.run took Params. However,
since a None query was used, these parameters were effectively ignored
(except for the timestamp). This commit makes the timestamps a
standalone argument that must always be passed when querying, and also
makes the getters take a Query instead of a Params. This allows much
finer control over what gets selected.

This change does affect a bunch of stuff further down. In particular,
find() no longer takes a Params, just a Query. The flow layer will take
care of filling in the query for each call (which was previously done in
ops). This adds a new trait, FillableQuery, which, well, indicates that
a query can be filled using a Params.