Add simple migration import
It is simple in the sense that it doesn't yet do the double-work
approach of our previous prototype in which an initializing node
continues to accept writes from upstream nodes, and then alternates
between receiving and processing once initialization has completed.
The feature is, however, complete, as far as I'm aware.

Note that this also introduces a TODO (which was an existing bug) around
detecting the initialization timestamp of new nodes that depend only on
new nodes. Those currently have their init_ts set to 0, which is
obviously incorrect.