NodeType enum for introspection on flow graph.
This change adds an enum that wraps the structs that implement NodeOp.
We need this in order to be able to introspect on the types of nodes in
the flow graph (petgraph requires all nodes to have the same type, and
we use the View trait) for transformations.

It's a little annoying that there is no cleaner solution to this, but
since Rust does not support downcasting, we're bound to using an enum
here.