Switch back from parking_lot to sync
Benchmarks suggests that sync:: locks are better at handling read-read
contention, yielding better read performance in the absence of writes.
It's slightly worse at read-write contention that parking_lot though, so
with the current benchmark setup, performance is likely to be lowered a
bit due to the high rate of writes. Delaying absorbs so that they are
buffered would also mitigate this problem.