partial fix for delayed update problem
it is only safe for nodes to expose state for timestamps they know they
will receive no more updates for. the concern is that a node may receive
additional records for time t from a different parent, which would
change the set of records a node should output for time t. if we expose
effects of updates in time t to queries while this can happen, the query
could see intermediate states, which is bad.

we used to implement this by not allowing a node to expose state from
time t until it receives an update from all parents with time t' > t.
specifically, queries only ever query absorbed state, and a node will
only ever absorb up to one less than the min timestamp of its children.
thus, while handling updates of time t, a node's parent would only serve
results from time t-1. it turns out that this is unnecessarily strict.

@jonathan observed that we can, in fact, safely expose data at time t
once we have handled *any* (non-None) update at time t. this is because,
a) when we start processing an update with time t, we know that all our
ancestors are up-to-date to at least t; and b) ancestor queries when
processing an update with time t are performed at time t. together,
these ensure that processing any single update at a given timestamp will
yield *all* records that would be produced at that time. thus, our state
is fully computed once that one update has been processed, and can
safely be exposed to queries.

now, it turns out that we can't quite get to this ideal yet. because
queries may only read absorbed state (for historical and performance
reasons), querying and absorption are linked. and while it *is* safe to
query a node at time t once it has processed an update at time t, it is
*not* safe for a node to immediately absorb updates after they have been
processed. the node's children could still want to query at t' < t. we
therefore still need the check that disallows absorbing past the min of
a node's children.

that said, the check can safely be modified (and is by this commit) to
call safe(min), instead of safe(min-1), since the child won't every
process a second update for any timestamp. this is not quite as good as
immediately exposing a processed update to queries though. to do that,
we would need to modify queries to read the current processed timestamp
of a node, and execute queries at that time, making them also scan parts
of the backlog. if this makes sense from a performance standpoint
remains to be seen. an alternative approach is to have nodes
periodically check the min of their children (and absorb if it has
changed), even if they have seen no updates. this feels less clean, but
would give better end-user query behavior.

as it stands, the system with this patch applied will expose slightly
more up to date data to queries, but not quite as up to date as they
could be. the issue of not seeing fully up-to-date views should be
somewhat mitigated, but is not removed altogether.

BTW: the patch is currently broken if you are using UNIONs. this is
because the forward path of UNIONs does *not* query into the other
ancestor views for other writes at the same time. thus those other
writes will *never* be output. this is terrible, and I don't know how to
fix it. pls hlp.