Rudimentary SQL -> FlowGraph translation mechanics
This adds a dependency on a (rudimentary) SQL parser built using nom,
and puts in place the basics of a pipeline that translates from SQL
queries into operator nodes. Only supports the most basic of queries at
this point, and requires an INSERT query for each write type to exist.