Eagerly push join columns through projections.
This is a hack to make naive building of Soup graphs from query graphs
work: even for filter/project nodes, we now push through all the base
table columns that subsequent join nodes will need. When establishing
the join conditions for the query graph edges, we hackily also add all
join columns to the column sets in the nodes on either side of the join.

The plan is to get rid of this hack and replace it with better logic
that only propagates the columns if needed.