Stopgap fixes to handle function columns
This fixes a bunch of errors related to functions on columns (as used
with aggregations) in the query graph/SQL incorporation logic. It's a
stopgap commit because we don't yet fix some issues "properly":

 1) Columns with functions applied are currently ignored when building
    the graph; we need to instead synthesize the correct operators from
    the "grouped" category.
 2) Anonymous function columns (i.e., those without an alias specified)
    are all called "anon_fn" currently, while we really want to give
    them unique generated names to disambiguate.

With these fixes, we can now incorporate all 28 TPC-W queries that we
can parse, albeit still without the aggregation nodes required.