Transactional bank transfer benchmark
This is a simple end-to-end test of transactions and also doubles as a
benchmark. It tries to repeatedly transfer money between two accounts by
(in a deliberately silly setup, to avoid genuine multithreading) first
reading two balances, then doing a transfer, and then tryint to do the
second transfer (based on a potentially stale read). This ought to
generate some aborts as well as successful commits.