Fix transactional_dispatch bug that erased updates
The loops over messages in `dispatch` and `transactional_dispatch`
were using the *same* key to insert into a map, thereby overwriting
earlier updates' entries with later ones if a node received a batch of
updates. Instead, we ought to concatenate the record vectors together,
so that all updates are preserved. This only affected transactions for
two reasons:
 1) `transactional_dispatch` was buggy; and
 2) the part of `dispatch` that handled the non-transactional case never
    returned any updates, so it being buggy did not get exposed unless
    it was called as part of a transactional dispatch.