Avoid join between votecount and rsum
This caused issues because of the lack of OUTER JOIN. The `total` view
would always be empty after a migration, since there would be no ratings
for the vote counts to join with. This also meant all subsequent views
were empty, which lead to overly fast migrations, and very fast empty
reads immediately following the migration.