Make RPC clients drive their own connection
It turns out that tarpc currently shares a *single* background thread
for driving all RPCs. See https://github.com/google/tarpc/issues/89, and
discussion at https://gitter.im/tarpc/Lobby?at=589245badcb66e4f76dbe4d6.
This in turn meant that all RPCs were effectively serialized, making all
performance results stupid.

This change instead drives the client in the same thread as the one that
does the RPC. This should fix the performance issue we've been seeing,
but also unfortunately currently breaks rustc:
https://gitter.im/tarpc/Lobby?at=58924b8e5309d6b358979df8
A fix is in the works
https://gitter.im/tarpc/Lobby?at=5892618a1e4d4bd962c1dfcf
But in the meantime, it is necessary to use @jonhoo's fork of tarpc
where this issue has a workaround applied.