Fix MySQL puts by using two connections
This fixes the problem described in
7f2ab4e3dfe203304f7e89be7b5c05942e74590d by using two connections in the
MySQL benchmarker, one for each statement that runs as part of `vote()`.
This is not ideal because it allows MySQL to parallelize the execution
of the `INSERT` and the `UPDATE` statement (unlike PostgreSQL, which
runs them one after another), but this will only inflate MySQL
performance.