Support recipe activation
This adds functionality that activates a recipe, adding its
expressions to an underlying Soup graph via the `SqlIncorporator`
mechanics. The idea is that the user modifies the recipe via calls
to `extend()` or `replace()` and finally calls `activate()` with
a `Migration` as an argument to actually compute and apply the
differences between the amended recipe and the underlying Soup graph.

Currently works for a trivial example, but crashes deep inside the
flow graph logic on a more involved migration; needs investigation.