Reuse if an identical query is added to the graph
This covers the (admittedly silly, but plausible in an ad-hoc query
setting) case that the a query that is identical to an existing one is
added.

Currently, this catches trivial re-writes too (e.g., different order of
projected columns). Once we have normalization, this may also catch
queries that are more expressed differently, but which are actually
semantically equivalent (and thus result in the same query graphs).