Full article data in memcached; handle all misses
We now cache the full article data in memcached in the "hybrid"
setup, rather than just vote counts, and also update the full entry on
every invalidation.

Also, this fixes a bug introduced by our hacky invalidation strategy:
since we temporarily write a NULL value with a 1-second expiration time
rather than deleting the entry (due to limitations in the rust-memcached
crate), we may occasionally read this value back. The code now correctly
handles this as a miss in memcached.