Ensure MSSQL reads hit the materialized view
In the non-enterprise version of SQL server, the "WITH (NOEXPAND)"
parameter is required to ensure that the query optimizer detects the
presence of a matching materialized view (the best "view matching"
algorithm is only available in the enterprise version). It's unclear
which version we get on Linux, but this makes sure we read from the
materialized view. The change also adds all other SQL server
configuration options required for materialized views to each
connection.

However, this does not seem to have had a substantial performance
effect (either on disk or on a ramdisk).