Use FutureClient/FutureService exclusively
Following https://github.com/google/tarpc/pull/121, using SyncClient
will start using an extra thread per client again, which we really don't
want. This change goes back to using a FutureClient with no magic
instead. It does rely on an unsafe impl Send for something containing a
reactor Core, which I *believe* is safe at the moment. Down the line
though, we'll want to monitor
https://github.com/tokio-rs/tokio-core/issues/179 and switch to a safe
exclusive core.

This change also fixes the broken server following
https://github.com/google/tarpc/pull/113. That PR changed the return
type of listen(), making one of them a Listen Future, which is *not*
automatically associated with the handle that is passed in. Thus,
running the core whose Handle was passed to `listen` is insufficient.
Don't know if this is a bug or not?