Must materialize between ingress and querying node
If there are nodes between a domain's ingress and a querying node (i.e.,
an operator that returns true from `will_query`), rather than the
ingress being a direct neighbor of the querying node, we were incorrectly
deciding to materialize nothing (not even the ingress).

Consider this setup:

 (ingress) -> (project) -> (join, QUERIES)

The old code neither propagated the querying obligation to the
projection (causing the ingress to be materialized), nor materialized
the projection. This change fixes this, using the following approach:

 1) If all nodes between a querying node and an ingress set
    `can_query_through` (e.g., a filter chain), we propagate the
    "inquisitive child" taint to the ingress's neighbors.
 2) If we hit a node that does not set `can_query_through`, we
    materialize its output.

Currently, only `Filter` nodes support `query_through`, but we can
implement it for projections too.