Expand implied table on `CREATE` and `INSERT INTO`
Previously, column names in the `CreateTable` and `Insert` SQL
statements would not store a table name. This was fine because it is
obvious: it's always the table that is currently being created. However,
to support primary keys more easily and to avoid ambiguities as to
whether `Column` structures processed by the SQL incorporation code
carry table names or not, we now also expand implied table names in
`CreateTable` and `Insert` statements.