Correctly set primary key for write types from SQL
When a `CREATE TABLE` SQL query specifies an explicit primary key, we
now set the same key (compound or plain) on the corresponding `Base`
node we crate for the query's write type. This makes sense because it is
fair to assume that future queries will use these key column(s), and
migrations to such queries will be more efficient if we've already got
an index on the relevant column(s).