Add schema versions to node names (WIP hack)
This incorporates schema versions to the names of base nodes as they are
created to avoid de-duplicating tables of the same name but different
definitions. Currently fails tests as queries aren't expressed in terms
of versioned tables. In the HotCRP migration benchmark, this leads to a
graph with 580 nodes at schema version 161.