Track query parameters, push their columns through
This change ensures that we don't accidentally project away those
columns which occur only as parameterized conjuncts of the WHERE
clause of a query (and not in the projected column set of the query).
Previously, we simply dropped such columns, and while we ended up
with the corrrect column set in the final leaf view, this view became
useless as it wasn't queryable because it lacked the parameter
column(s), which we use as keys. Now we push such columns all the
way through by including them in the query graph nodes' column sets.