Set correct key column (query param) on leaf views
This fixes a long-standing issue with our SQL-based query incorporation:
the `Reader` nodes corresponding to each `SELECT` query did not have the
right key columns assigned because we lacked the information about query
paramters and didn't push parameter columns through to the leaves. Now
we assign the right key, although we cannot handle compound keys yet
(they need support in `evmap` and in the `Reader` API). For queries
without parameters, we still default to the first column.