Re-use existing query if only parameters differ
This handles the best-case for Soup re-use: we already have the exact
query we need, and the user simply asked to add it again,
parameterized on another column. We can reuse the entirety of our
data-flow graph and simply need to add a new reader node. Since Soup
only allows one `Reader` child per node (currently), we also add an
identity node to hang the new reader off. The cost of a migration
that performs this re-use is the cost of building a new reader
materialization via replay from the closest materializations.