Restored index push-up on migration
This fixes a bug discovered by @larat: when we introduced multi-key and
compound key support, we lost the code to push indexing requirements
upwards from a querying node to the closest materialized ancestors. This
wasn't a problem until we had a transitive need to query an index
through an operator:

 `Join` -> `Filter` -> `Ingress`

Since `Filter` does not set `will_query` (since it does not query out of
its own volition), the materialization on the `Ingress` node -- which as
of the fix in 42406b8c16e9a955f33c37fdd5d1054a36c41635 was correctly
established -- ended up being removed as unnecessary as it never found
out about the correct index key to use.

This change brings back index obligation push-up and adapts it for
multi-key views and compound keys.