Populate HotCRP with some test data
This test data is for the latest schema version (161), so we push it in
after migrating through all the previous schemas. Note that some binary
fields must be re-encoded as hexademical text in the dataset since Soup
does not support binary fields. This only affects the recently changed
`sha1` fields.