Crude query extraction from HotCRP code base
Adds a utility that walks a HotCRP checkout and extracts anything that
look like a SELECT query. We try (not very hard) to make the queries
more digestible and then dump the ones that we understood into a recipe
file. Currently, this manages to extract between 1/3 and 1/4 of
query-like strings in the code into correct SQL queries (17/82 at schema
v1, 82/165 at schema v161).