Ambiguous columns are sometimes ok, pick randomly
Queries with joins often refer to tables that have columns of identical
names, in which case we'll see ambiguity in resolving columns in the
output (even if no such ambiguity exists because the join would only
have projected one of the join columns). As a workaround for a more
comprehensive solution, choose a random column.

Note that this can, unfortunately, mask invalid queries when there is
genuine ambiguity (i.e., the columns aren't known to be the same because
they're joined upon).