Initial work on MIR
This starts decoupling the high-level SQL/`QueryGraph` logic from the
low-level Soup graph via an intermediate MIR. The MIR is similar to the
Soup graph (its node structures hold all the constructor arguments
that will be required to instantiate the corresponding Soup nodes), but
can be modified after construction (e.g., rewritten as optimizations are
applied).

This commits adds the foundations of the MIR, and ports `Base` node
creation to use the MIR rather than changing the Soup graph directly.