Schema versions in SQL and MIR; base node reuse
More progress towards a working MIR layer: we now represent schema
versions in the SQL and MIR layers, and use them when looking for
opportunities to reuse base nodes with unchanged schemas. This also gets
us back to where we were before the MIR layer addition as far as
incorporating base nodes is concerned: we can run through all 161 HotCRP
schema versions and end up with ~280 base nodes with unique schemas;
currently, we always add a new base node if the schema differs, but this
will change soon.