Start bringing queries back; separate QG/MIR reuse
This brings back the core of the query graph-based high-level reuse
we've used with read queries, and ties it into the new MIR-based design.
We compute the QG and try high-level reuse strategies before we compute
the MIR for a query (and thus also before any MIR-level optimization or
reuse happen, since these are more costly than high-level query reuse).
Compiles, but crashes on the first query, so clearly buggy!