Don't unnecessarily recompute columns in MIR->flow
`MirNode`s already carry the correct set of fiinal columns as a field,
so there's no need to recompute it in an operator-specific way for each
implementation of `into_flow_parts`. The duplication was a result of me
reusing too much old code while porting, and gave rise to several bugs.