Re-register reused nodes at current schema version
Previously, we would only register MIR nodes in the `nodes`
structure of the `SqlToMirConverter`when they are first created.
However, this meant that queries failed to find (correctly) reused base
nodes after a schema version upgrade, so we now re-register nodes on
reuse for the current schema version. This ensures that there is a
registered node for each schema version in which this node exists
(either as a base or as part of a query).