Apply alias on computed columns in grouped nodes
Normally, we only apply aliases at the leaf of a query, but the computed
columns in grouped operators (e.g., aggregations) are an exception:
because these are new columns that are (a) introduced by the operator
and (b) not identical to any parent columns or base table columns, we
can rename them to their aliases immediately. This fixes the tests
broken accidentally by bc71c1960fa4cd5232015d24d287ca6c4eaed113.