Correctly handle lookups into Target nodes
Consider the classic L*R -> J -> R' join example where L and R are
partially materialized, and R' is a reader. L gets a message with key
K1, misses in its output state, and thus does a replay of K1 all of its
own volition. It is not in the Target state. R' then requests a replay
of key K2, which is replayed through J from R. The replay for K2 also
misses in L when it passes through J. Previously, J would observe that L
was already in the Target state, and dutifully add itself to the
"queued" list of L. In contrast, if the R' replay happened first, it
would switch L to Target, and then add itself as "release" on L.

The difference between these two was that the buffer of the "release"
node is replayed first, then the buffer of the Target node, then the
buffer of any "queued" nodes. Unfortunately, when we get this particular
scenario where J ends up in "queued", an L update can be propagated to J
while L still has more updates buffered (as discussed in the huge
comment in this commit). If those then miss in R's buffer, we have an
issue, since we both need to make R a Target, *and* keep replaying L's
buffer.