Also pull down referenced but unprojected columns
Some operators refer to parent columns in their implementation, but do
not project these columns into their output: consider a filter on a
column that is subsequently discarded, or the column that a computed
column (aggregation, min/max) is computed over. We now correctly pull
these column from ancestors in the MIR rewrite phase.