Extract join tables from "ON" in explicit join
As noted in ae81bcf69820ec7a1d36d5c69d193c20f7dc3780, we must actually
determine the left and right tables of each explicit join by analysing
its join predicate. This change introduces code in the QueryGraph
implementation do to so. However, it unfortunately still has to maintain
the hacky logic that flips the ConditionTree in the predicate according
to which column ends up being the left one.