Be more economical with MySQL connections
By default, `mysql::Pool` creates a pool with a minimum of 10
connections (and a maximum of 100). When we run with many clients in
separate processes, each of our clients only uses one or two
connections, but we still create 10. This quickly leads to the server
running out of connections when faced with larger numbers of clients. We
force the pool to have between 1 and 4 connections to avoid this.