Rate-limit vnf client to avoid spamming replay
We believe this is okay, because a "real" system would have clients
block, and use many clients to get continued read performance.
Fundamentally, the system cannot support a replay rate greater than its
write rate, which is what this sleep ensures. Another possible fix to
this would be to expose a bounded channel to the Reader, and then
try_send, returning E_BUSY it it was full. This would have the same
effect, though at some more implementation complexity.