Distinguish between dropped and captured
The previous code didn't work correctly, because m can also be
Packet::None if it hits a Reader. In that case, we would still break
'outer, which for full materialization caused us to not set `finished`
below.

We also can't just check whether `i` is the last node in the path.
Imagine what would happen if the output of a union was partially
materialized with the previous code. Would it it break 'outer?