Hacky patch for better domain parallelism
Applies @fintelia's close-to-deadline patch to make domains spin
on their input channels for 10ms rather than doing blocking `select`
(which turns out to have very slow wakeups); this achieves much better
performance for parallel domains and allows better evaluation of
different domain setups.

We'll probably want to merge a version of this into `master` eventually;
for now, this records the patch and makes it available to @omegablitz
for use with domain assignment algorithm benchmarks.