Replay through base nodes should NOT materialize
This fixes a subtle bug that had major correctness implications. When
the add/remove column optimization was introduced to base nodes
(enabling a base to be reused if a schema migration only adds or removes
a column), we decided to replay a base node through *itself* in order to
adapt the record format to the new column set and add default values to
old records. This is necessary because the replay destination node is
new, so it may use new columns records.

Replaying a base through itself, however, would cause all records that
are replayed to be materialized, *doubling* the size of the base table
every time a replay was run for it! This not only led to incorrect query
results, but also (likely) caused the memory blowup in the HotCRP
migration benchmark.

The solution is to avoid materialization in the case when the current
node is a base table *and* the incoming records are part of a replay,
but still materialize otherwise.