Have left joins avoid emitting positive/negative null pairs
Still room for improvement. Currently just catches the case where the number of
nulls for a given key should stay the same, and prevents any positive or
negative nulls for being emitted. However, it still will emit 2 positive and 1
negative null, when it really should just emit a single positive.