Important FIXME note
It's not clear to me what the best way to apply this is, though. The
unique column ID is only known inside the `adapt_base_node` method, but
must be passed back up and stored in the MIR representation, so that
later removal of the column knows what index to use. In many ways, it
would be preferable if the base node implementation did the index
remapping internally...