Absolute column ID resolution on base nodes
Adds a `column_id_for_column()` method to `MirNode`. On a base node,
this resolves the absolute ID of the column passed in the argument,
subject to column additions and removals in migrations using the
`ColumnSpecification` held by the base MIR node. On non-base nodes,
it simply identifies the offset of the column in the node's named column
vector.

This fixes correctness for nodes that directly follow base nodes, and
also has the nice side effect of simplifying the node generation code by
replacing a commonly repeated `.iter().position(...).unwrap()` idiom.