Smarter `COUNT(*)` rewrite hack
Previously, the rewrite pass would pick the first or last column of the
table. We now pick a column not mentioned in the GROUP BY clause or the
WHERE clause to avoid grouping by the aggregation column. The heuristic
is currently a bit overeager: implicit joins, whose columns are
mentioned in the WHERE clause, blacklist their columns for the rewrite
pass, but this isn't a problem in practice, since any column will do and
there is usually at least one column not mentioned in either clause.

This allows us to go all the way through HotCRP up to v161 with column
add/remove on base nodes, although there are some runtime errors related
to sharders.