Revamp replay logic for sharded graphs
This change is fairly finicky, and so warrants some further discussion.

First, let's do a recap of how replay in distributary generally works:

Whenever the system decides (during migration) that a node should be
materialized, it finds a set of paths that need to be replayed to fill
this new materialization. There will be more than one path for any given
materialization only in the case of unions (joins only need one side to
be replayed). Each such path is given a /tag/, which uniquely identifies
that replay path. Paths never include materializations -- they are a
path from the target materialization to its *nearest* ancestor
materialization.

The system decides at migration time whether the replay should be "full"
or "partial". To trigger a full replay for a path, the migration code
first informs every domain on the path about that replay path, and then
prompts the domain responsible for the first node on the path to start
replay. Every message ("replay piece") that is a part of that replay is
marked with the path's tag, and the last piece has a flag to indicate
that the replay is done. For a partial replay, the process is similar,
except that the migration thread does not trigger a replay from the
source domain. Instead, it sets up a connection between the source and
the destination domain of the path, such that the destination domain can
send a control message to the source domain when it needs a particular
key replayed. When such a request arrives, the source domain just sends
along the state for that key (with the appropriate tag), which will
eventually reach the destination domain.

The part where this all gets complex is when the graph forks and joins.
In a naive implementation, if a replay piece travels across a fork-join,
the destination domain would receive multiple "parts" for each replay
piece (e.g., for a single key in the partial replay case). This would be
unfortunate, as the domain might expose partially replayed state. To
circumvent this, unions are "special" in that when they hear a replay
piece from one ancestor, they will *wait* until they also hear a replay
piece from every other ancestor, and then release all the pieces as a
single piece.

Sharding complicates this picture, because when we de-shard, there is
still conceptually just *one* path. We could place an unsharded union
after the sharded part of the graph, but the union wouldn't know how
many pieces to wait for for any given replay (because it doesn't know
how many shards were involved).

This commit modifies replay piece messages such that they include how
many shards they are present on. If a replay is requested *from* a
sharded node, this is set to the number of shards, whereas if the replay
originates *above* some sharded subtree, the sharder node sets the
number to the number of shards it ended up sending to. This way, the
union can look at the number included in the replay piece to see how
many pieces it should wait for before releasing.

To enable the union to see the replay pieces coming from different
shards as separate, a little bit of hackery was required. Since sharding
is transparent to the graph structure, the union is really only
receiving updates from "one" node, and thus it'd be tricky for it to
differentiate between the pieces from different shards. We get around
this by having sharded Egress nodes set their source address as the
shard id instead of their node ID, and having Ingress nodes above
de-sharding unions not overwrite that value. It's a bit of a hack, but
it works.

Fun fact: while writing I also realized that there is *still* a bug in
the current implementation. Since we don't require a de-shard for a
shuffle, the sharded Sharder nodes in a source domain would all set
*different* values for the number of shard pieces for the piece they
forward, making the union very confused.