Project literal columns into edge views
Adds the final plumbing required to project literal values specified in
the field list of a SQL query into the corresponding edge view. We
project the columns close to the edge view as they do not need to ever
be computed on and might complicate re-use if added earlier.

Also fixes some compiler warnings.