Start work on distributed Soup daemon binary
This puts in place the skeleton of a daemon ("gulaschkanone") for
distributing Soup. Currently, it does nothing other than listening for
connections (if in "controller" mode), or connecting to a controller
(if in "worker" mode). The plan is to have workers register and
regularly heartbeat the controller, and for the controller to assign new
domains to the workers to execute.