Correctly reuse full query with additional params
If we have the opportunity to reuse an entire existing query because a
new query is identical, but merely has different parameter(s), we may
need to base the reuse off a node that isn't the direct ancestor of the
existing leaf. The reason is this: the final data-flow node (i.e.,
non-reader node) in a query is usually a projection that drops columns
we don't need (but preserve up to the end anyway to maximize reuse
opportunities). However, the reusing query may have additional
parameters on columns that were projected away for the old reader
because they aren't listed as projected columns in the query. Hence, we
backtrack until we find a node that has all the right columns, and base
ourselves off that.

Fixes a bug reported by @larat7, but currently has two restrictions:
 1) We don't add the potentially necessarry projection to the new,
    extended query (and we may thus end up with too many columns).
 2) We could, in theory, backtrack over non-projection nodes and end up
    with a prefix of the query only, although this ought not to happen
    because we're already by definition in a situation where we can
    reuse the entire query minus the final projection. Might want an
    assertions nevertheless, though ;-)