Also register with local channel coordinator
Before informing the controller about a booted domain, we have to
register it with the local channel coordinator (the bootup code does not
itself do this, but I had assumed it did). In the absence of this,
writes will fail because domains aren't connected.

Distributed "vote" now 'works', modulo quirks related to readers.