Fix reprojection on ReaderOntoExisting reuse
This fixes a bug that occured if we chose `ReaderOntoExisting`-style
reuse with an existing query: we concluded that the query is identical,
except that the free parameter is on a different column. However, this
previously non-parameter column may well have been projected away in the
original query; hence we backtrack to the last node that has the columns
needed (see 34f98e9f). Since this is above the original projection, the
new query now needs to add a new projection node, rather than simply an
identity node on whatever node we've branched off from, to avoid
including spurious columns in the output. We now have a test for this,
too.