Differentiate lookup and non-lookup obligations
Specifically, some nodes (Readers in particular) do not perform lookups
into their ancestors, and thus do not require them to be
materialized/query-through, but would still benefit from an ancestor
that has an index on their key so that they can be partially
materialized.

This commit only enables this behavior for Readers, but we could enable
it for more operators (such as aggregations) down the line.
Alternatively, we could make it such that lookup keys automatically also
propagate as "replay keys" to ancestors. Topic for debate.